Revert "new constructor for Item and randomizing armorValue for armors"
This reverts commit 75abef6974fef3acaa3ff36ff4f23f54592e32d2.
