+++ RogueClone/RogueClone/RogueClone.csproj
@@ -85,13 +85,13 @@
     <Compile Include="Stats\Health.cs" />
     <Compile Include="Charaters\Hero.cs" />
     <Compile Include="Items\Item.cs" />
-    <Compile Include="Items\Defense\WizardArmor.cs" />
+    <Compile Include="Items\Defense\LeatherArmor.cs" />
     <Compile Include="Stats\Level.cs" />
     <Compile Include="Items\Offense\WizardWeapon.cs" />
     <Compile Include="Stats\Mana.cs" />
     <Compile Include="NPCs\MonsterTierTwo.cs" />
     <Compile Include="NPCs\MosterTierOne.cs" />
-    <Compile Include="Items\Defense\RogueArmor.cs" />
+    <Compile Include="Items\Defense\PlateArmor.cs" />
     <Compile Include="GameMain.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Charaters\Rogue.cs" />
